fix test by re-init data on each sim
When started this unit test, I had only one simulation performed, and
the data structures were set up in the setup_method(), but then for the
optimization tests, multiple simulations were run.  Instead of setting
up the data structures once, we needed it prior to every run.

With that in place, the tests now pass 👌