add example autograd member primitive
TVB implements most arithmetic in class members, while autograd is
expecting free standing functions.  Forunately, this can be resolved
by delaying primitive & vjp registration after instantiation of the
object.

This is doubly useful because it will ensure that we can specialize the
vjp operator for the runtime use case and not rely on general cases.
When most parameters are constant, this will greatly simplify the
calculation of vjp as those constants do not need to be handled.

Specializing on a per object basis may bloat autograd's registry a bit
but ensures that if the model is instantiated more than once, the
appropriate vjp will be generated and used.

This commit just adds a test case showing how this can be done.