first sketch of symbolic jacobians
symbolic jacobians are in reach for most of our models,
especially with the upcoming DSL.  this is a first
sketch which computes derivatives symbolically with
sympy for Jansen Rit and generates a sequences of
instructions which can be strung together, eval'd and
then jit'd with numba.

one could then also specialize when parameter values
are known by substituting the values in prior to
doing the symbolics, which will help as well.