coupling template w/ unit and integration tests
Monolithic templates are hard to debug, while small templates don't
compose well because the template engine doesn't understand data flow
and semantics in the generate code.  The compromise, which this commit
adds an example of, is to have the units of templates correspond to
functions in the target language, with a unit test verifying the
function in isolation (with  a trivial wrapper kernel), and an
integration test which uses in function in a full simulation kernel.